% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immosaic.R
\name{immosaic}
\alias{immosaic}
\title{Make a packed image mosaic}
\usage{
immosaic(
  im,
  width = 1024,
  height = 800,
  mask = NULL,
  weights = NULL,
  preferred = NULL,
  max_num_tries = 100,
  scales = c(rep(0.5, 2), rep(0.25, 4), rep(0.15, 8)),
  scale_fun = function(i, s, c) {     if (c < (i * 0.5)) {         mscale <- min(s)    
        c(s, rep(mscale/2, floor(1/mscale)))     }     else {         scales     } },
  max_images = 2000,
  min_scale = 0.05,
  bg = "transparent",
  show_every = 25,
  ...
)
}
\arguments{
\item{im}{Can be either a character vector of image
file names (format must be compatible with \code{\link[imager:load.image]{imager::load.image()}}),
a list of \code{\link[imager:cimg]{imager::cimg}} objects, or a function that generates
an image when evaluated. Tje function can take a single argument,
which is the current iteration of the packing algorithm. Can also
be specified as an \code{rlang} style lambda syntax (see \code{\link[rlang:as_function]{rlang::as_function()}}).}

\item{width}{Width in pixels of produced image}

\item{height}{Height in pixels of produced image}

\item{mask}{An optional masking image.}

\item{scale_fun}{}
}
\value{

}
\description{
A simple greedy algorithm tries to pack as many images
into a larger image as possible, taking into account
transparency, if available (recommended).
}
