% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impac-package.R
\docType{package}
\name{impac-package}
\alias{impac-package}
\alias{_PACKAGE}
\title{impac: Make packed image mosaics}
\description{
impac takes a list of images or a function that generates images and packs them onto a canvas using a greedy algorithm, while respecting transparency.
}
\author{
\strong{Maintainer}: Russell Dinnage \email{r.dinnage@gmail.com} (\href{https://orcid.org/0000-0003-0846-2819}{ORCID}) [copyright holder]

}
\keyword{internal}
