% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip.R
\name{impac_clear_cache}
\alias{impac_clear_cache}
\title{Clear any cached \code{impac} objects}
\usage{
impac_clear_cache()
}
\value{
No return value
}
\description{
Clear any cached \code{impac} objects
}
\examples{
impac_clear_cache()
}
