% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip.R
\name{impac_resume}
\alias{impac_resume}
\title{Function to resume an image packing where it left off}
\usage{
impac_resume(x = NULL, ...)
}
\arguments{
\item{x}{An \code{impac} object created from a previous run of \code{impac}.
Can also be left blank in which case this function attempts to recover
the latest \code{impac} run from the cache (see \code{\link[=impac_recover]{impac_recover()}} for details).}

\item{...}{Other arguments to be passed to \code{\link[=impac]{impac()}}. By default, original
arguments from the original call used to make \code{x} will be used. Passing an
argument here will override the original arguments.}
}
\value{
An \code{impac} object
}
\description{
Function to resume an image packing where it left off
}
